Migrate getIn / hasIn
