+++ src/utils/coerceKeyPath.ts
@@ -1,7 +1,10 @@
+import type { OrderedCollection } from '../../type-definitions/immutable';
 import { isOrdered } from '../predicates/isOrdered';
 import isArrayLike from './isArrayLike';
 
-export default function coerceKeyPath<I>(keyPath: Iterable<I>): Iterable<I> {
+export default function coerceKeyPath<I>(
+  keyPath: OrderedCollection<I> | ArrayLike<I>
+): ArrayLike<I> {
   if (isArrayLike(keyPath) && typeof keyPath !== 'string') {
     return keyPath;
   }
